# Speedup `brew --version`.