--- a/Library/Homebrew/test/cmd/--caskroom_spec.rb
+++ b/Library/Homebrew/test/cmd/--caskroom_spec.rb
@@ -1,23 +1,19 @@
 # typed: false
 # frozen_string_literal: true
 
-describe "brew --caskroom", :integration_test do
-  let(:local_transmission) {
-    Cask::CaskLoader.load(cask_path("local-transmission"))
-  }
+require "cmd/shared_examples/args_parse"
 
-  let(:local_caffeine) {
-    Cask::CaskLoader.load(cask_path("local-caffeine"))
-  }
+describe "brew --caskroom" do
+  it_behaves_like "parseable arguments"
 
-  it "outputs Homebrew's caskroom" do
-    expect { brew "--caskroom" }
-      .to output("#{HOMEBREW_PREFIX/"Caskroom"}\n").to_stdout
+  it "prints Homebrew's Caskroom", :integration_test do
+    expect { brew_sh "--caskroom" }
+      .to output("#{ENV["HOMEBREW_PREFIX"]}/Caskroom\n").to_stdout
       .and not_to_output.to_stderr
       .and be_a_success
   end
 
-  it "outputs the caskroom path of casks" do
+  it "prints the Caskroom for Casks", :integration_test do
     expect { brew "--caskroom", cask_path("local-transmission"), cask_path("local-caffeine") }
       .to output("#{HOMEBREW_PREFIX/"Caskroom"/"local-transmission"}\n" \
                  "#{HOMEBREW_PREFIX/"Caskroom"/"local-caffeine\n"}").to_stdout