--- a/Library/Homebrew/test/cmd/--version_spec.rb
+++ b/Library/Homebrew/test/cmd/--version_spec.rb
@@ -4,10 +4,8 @@
 require "cmd/shared_examples/args_parse"
 
 describe "brew --version" do
-  it_behaves_like "parseable arguments"
-
-  it "prints the Homebrew version", :integration_test do
-    expect { brew "--version" }
+  it "prints the Homebrew's version", :integration_test do
+    expect { brew_sh "--version" }
       .to output(/^Homebrew #{Regexp.escape(HOMEBREW_VERSION)}\n/o).to_stdout
       .and not_to_output.to_stderr
       .and be_a_success