--- a/Library/Homebrew/test/tap_spec.rb
+++ b/Library/Homebrew/test/tap_spec.rb
@@ -210,9 +210,7 @@ describe Tap do
     setup_git_repo
 
     expect(homebrew_foo_tap.git_head).to eq("0453e16c8e3fac73104da50927a86221ca0740c2")
-    expect(homebrew_foo_tap.git_short_head).to eq("0453")
     expect(homebrew_foo_tap.git_last_commit).to match(/\A\d+ .+ ago\Z/)
-    expect(homebrew_foo_tap.git_last_commit_date).to eq("2017-01-22")
   end
 
   specify "#private?" do