… Library/Homebrew/cmd/--version.rb